Added delayed jobs. You can specify at what time or how far into the future to run a job. It's also used to delay retries when a job fails.
Currently it's implemented using an extra list that gets SORTed and checked on every pop. Redis 1.1 will have sorted sets which will make it easier. However, what's really needed is a sorted queue. Might need to put a request in a for a ZPop and see if others would find it useful.